build: move ingestion out of conversion +
- move ingestion out of conversion
- only ingest supported content types

Ref: metriport/metriport-internal#1050
