+++ packages/api/src/external/commonwell/document/document-query.ts
@@ -594,6 +594,9 @@ export async function downloadDocsAndUpsertFHIR({
             }
           }
 
+          // ingest the document into OpenSearch
+          await ingestIntoSearchEngine(patient, file, requestId);
+
           const FHIRDocRef = toFHIRDocRef(doc.id, docWithFile, patient);
 
           if (shouldUpsertFHIR) {
