+++ packages/api/src/external/fhir-converter/converter.ts
@@ -5,7 +5,6 @@ import { Util } from "../../shared/util";
 import { sandboxSleepTime } from "../commonwell/document/shared";
 import { makeFHIRServerConnector } from "../fhir/connector/connector-factory";
 import { buildDocIdFHIRExtension } from "../fhir/shared/extensions/doc-id-extension";
-import { makeSearchConnector } from "../opensearch/connector-factory";
 import { sidechainConvertCDAToFHIR } from "../sidechain-fhir-converter/converter";
 import { FHIRConverterSourceDataType } from "./connector";
 import { makeFHIRConverterConnector } from "./connector-factory";
@@ -74,16 +73,6 @@ export async function convertCDAToFHIR(params: {
       patientId: patient.id,
       documentId: document.id,
       payload: JSON.stringify({ s3FileName: jsonFileName, s3BucketName }),
-      requestId,
-    });
-
-    // ingest the document into OpenSearch
-    const openSearch = makeSearchConnector();
-    await openSearch.ingest({
-      cxId: patient.cxId,
-      patientId: patient.id,
-      s3FileName,
-      s3BucketName,
       requestId,
     });
